pyproject.toml
[project]
name = "three-pigs-function-calling"
version = "0.1.0"
description = "OpenAI Function Calling Demo using the Three Little Pigs tale"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "sympy>=1.10.0",
]
